Backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
mysql-connector-python==9.0.0
pydantic==2.9.2

